Carpanta::Commands::CreateAppointment completed
