+++ config/initializers/commands.rb
@@ -1,0 +1,6 @@
+require 'app/commands/create_appointment'
+
+Carpanta::Commands::CreateAppointment.configure do |config|
+  require 'domain/appointments/repository'
+  config.repository = Carpanta::Domain::Appointments::Repository
+end
